// shared/types/attractions.ts
export type Attraction = {
    m_id: number;
    title: string;
    other_titles?: string;
    type?: string;
    idea_author?: string;
    architector?: string;
    sculptor?: string;
    year_arise?: number;
    country?: string;
    city: string;
    address?: string;
    description?: string;
    history?: string;
    trivia?: string;
    yandex_link?: string;
};